Allow handles to store native fonts and return those from 'Handle::load' if possible.
This fixes another chokepoint in Iosevka loading, as we were needlessly reading fonts from disk and stuff with load, whereas we've already had a font at our disposal before creating a handle. That seemed wasteful.
Note that I did cargo check --all-features this code, but I am only sure of coretext part compiling. If we pursue cross-platform support, we'll need to revisit this (probably). The idea from this commit is not specific to this commit though.