Faster OTC font loading (#1)
This commit vastly simplifies the way .otc parsing is done in font-kit by initializing a font from a descriptor and not copying the buffer. It also makes Handle capable of storing an already loaded font, as what happens currently is that we first have a Font which we turn into a Handle (losing track of the Font that was used to create it) only to recreate a font from that same handle (and reparsing the font file in the process). Less waste is good.

I've had another stab at it over the weekend at https://github.com/osiewicz/font-kit/pull/new/improve_font_parsing_perf but changes from this PR are way smaller and make more sense IMHO. improve_font_parsing_perf tried to deal with issues in font-kit's implementation of otc parsing (it had an accidental O(n^2)), but it couldn't deal with the fact that font-kit had to allocate about 16Gb of memory (kid you not) to parse font-kit file. :))))) Each 300Mb allocation took about 30ms, and we had to make 54 of those. Quick math: by allocations alone that approach couldn't get below 1.5s for parsing Iosevka, no matter how hard we tried.

Meanwhile, on this branch parsing Iosevka takes about 30ms and most of that time is spent.. well.. parsing the font I guess?